Build RPM packages using a spec file
Drop the use of fpm

Also build and link on same arch,
before we built on buildarch and linked on targetarch, but because we
use depot.dev with native machines for building multiarch images that's
not needed any more.

Use the default RPM CFLAGS and LDFLAGS when building and linking the
binaries/libraries.